requirements.txt
pyTelegramBotAPI==4.14.0
yt-dlp==2024.04.09
requests==2.31.0
python-dotenv==1.0.1
